refactor: Complete Zustand store modularization
✅ REFACTORING COMPLETE

Changes:
1. Created itinerarySlice.ts (335 lines)
   - Itinerary editing (CRUD operations)
   - Planning phase management
   - Requirements checklist
   - Auto-progress functionality

2. Integrated all slices into new useStore.ts
   - chatSlice (45 lines)
   - uiSlice (85 lines)
   - toastSlice (36 lines)
   - settingsSlice (132 lines)
   - historySlice (101 lines)
   - itinerarySlice (335 lines)
   Total: 734 lines (vs original 766 lines)

3. Fixed type exports for backward compatibility

Benefits:
- Clear separation of concerns
- Each slice is independently testable
- Improved maintainability
- No breaking changes to existing API

Test Results:
- ✅ 76/76 unit tests passing
- ✅ All utility functions tested
- ✅ Type-safe implementation

Note: Existing errors in QuickActions.tsx are pre-existing
and unrelated to this refactoring.
